Add isRooted property to nodes
Previously, I added a "rootedNodes" array property to the branch
thinking that I will use this upon submission to determine which
parts of the branch are rooted to the universal tree and which are not.
But upon further thinking, doing this would require traversing the whole
branch and more complex calculations which will probably affect
performance. Having a boolean on each node however makes things simple:
all nodes that have isRooted === true would be part of the rooted branch
and the rest would be part of the hanging branch.