Implement Skill Node validation
For now, skill node validation merely consists of informing the user
about disconnected skill nodes which instead of being added to the
main tree display will be added to a yet-to-be-implemented Requests
page. Requests are a collection of skill nodes or branches that we
want to be connected to the main tree via modules that some more
knowledgeable user could make for us.